chore: add Field & mysql's my_bitmap class

+++ include/constants.h
@@ -118,7 +118,7 @@ struct MYSQL_LEX_CSTRING {
 typedef struct MYSQL_LEX_CSTRING LEX_CSTRING;
 
 constexpr int INT_OFFSET = 4;
-constexpr int SQL_SIZE_ARRAY[] = {248,  744,  3304, 3208, 3208,
+constexpr size_t SQL_SIZE_ARRAY[] = {248,  744,  3304, 3208, 3208,
                                   2040, 2040, 1968, 264,  224};
 
 } // namespace loft
